requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.86.0
plyer==2.1.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.1
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
tornado==6.5.1
tqdm==4.67.1
tradingview-ta==3.3.0
twilio==9.6.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
